Refactor state to use localStorage and re-arrange src directory
