+++ src/components/classBased/components/todoContainer.js
@@ -1,28 +1,29 @@
 import React from 'react';
-import Header from './header';
+import Header from '../../functionBased/components/header';
 import InputTodo from './inputTodo';
 import TodosList from './todosList';
 import { v4 as uuidv4 } from 'uuid';
 
 class TodoContainer extends React.Component {
   state = {
-    todos: [
-      {
-        id: uuidv4(),
-        title: "Setup development environment",
-        completed: true
-      },
-      {
-        id: uuidv4(),
-        title: "Develop website and add content",
-        completed: false
-      },
-      {
-        id: uuidv4(),
-        title: "Deploy to live server",
-        completed: false
-      }
-    ]
+    todos: []
+  }
+
+  componentDidMount() {
+    const temp = localStorage.getItem("todos")
+    const loadedTodos = JSON.parse(temp)
+    if (loadedTodos) {
+      this.setState({
+        todos: loadedTodos
+      })
+    }
+  }
+
+  componentDidUpdate(prevProps, prevState) {
+    if(prevState.todos !== this.state.todos) {
+      const temp = JSON.stringify(this.state.todos)
+      localStorage.setItem("todos", temp)
+    }
   }
 
   handleChange = (id) => {
