+++ src/components/classBased/components/todoItem.js
@@ -4,6 +4,10 @@ import styles from './styles/TodoItem.module.css'
 class TodoItem extends React.Component {
 
   state = { editing: false }
+
+  componentWillUnmount() {
+    console.log("Cleaning up...")
+  }
   
   handleEditing = () => {
     this.setState({
